refactor(python-sdk): remove parse_* wrapper functions and consolidate prompt rules
- Remove deprecated parse_elemental_response and parse_express_response wrapper functions.
- Update all integration and compiler tests to instantiate and call ElementalParser and ExpressParser directly.
- Move format-specific rules (constants) and transform_examples helper methods into ElementalPromptGenerator and ExpressPromptGenerator.
- Remove format_description from InferenceFormat base class and subclasses.
- Retain catalog_description on the strategy classes (ElementalFormat, ExpressFormat, TransportFormat) as requested.
