+++ agent_sdks/python/a2ui_agent/src/a2ui/inference_formats/experimental/elemental/parser.py
@@ -19,92 +19,95 @@ from typing import Any, List, Union
 from a2ui.core.catalog import Catalog
 from a2ui.schema.catalog import A2uiCatalog
 from a2ui.parser.response_part import ResponsePart
-from .compiler import ElementalCompiler
+from a2ui.parser.parser import Parser
+from google.adk.utils.feature_decorator import experimental
+from .compiler import ElementalCompiler, TAG_PREFIX
 
 _A2UI_OPEN_PATTERN = re.compile(r"<a2ui\b[^>]*>", re.IGNORECASE)
 
+_BLOCK_PATTERN = re.compile(
+    r"<a2ui\b[^>]*>.*</a2ui>"
+    f"|<{TAG_PREFIX}delete-surface\\b[^>]*>(?:.*?</{TAG_PREFIX}delete-surface>|/>)?"
+    f"|<{TAG_PREFIX}call-function\\b[^>]*>(?:.*?</{TAG_PREFIX}call-function>|/>)?",
+    re.DOTALL | re.IGNORECASE,
+)
 
-def parse_elemental_response(
-    content: str,
-    catalog: Union[Catalog[Any, Any], A2uiCatalog],
-    surface_id: str = "main",
-) -> List[ResponsePart]:
-    """Parses response containing A2UI Elemental HTML and compiles it to ResponseParts.
 
-    NOTE: This parser supports unclosed tag auto-closing for real-time streaming preview
-    rendering. If the final <a2ui> block is unclosed (truncated), it will be auto-closed
-    and compiled with is_final=False to discard any trailing incomplete statements.
+@experimental
+class ElementalParser(Parser):
+    """Concrete parser implementation for A2UI Elemental TSX/HTML5 responses."""
 
-    Args:
-        content: The raw LLM response.
-        catalog: A Catalog or an A2uiCatalog.
-        surface_id: The target surface ID.
+    def __init__(self, catalog: Union[Catalog[Any, Any], A2uiCatalog], surface_id: str = "main"):
+        self.catalog = catalog
+        self.surface_id = surface_id
+        self._truncated_blocks = set()
 
-    Returns:
-        A list of ResponsePart objects containing compiled JSON payload list.
-    """
-    content_lower = content.lower()
-    last_open_match = list(_A2UI_OPEN_PATTERN.finditer(content))
-    last_close = content_lower.rfind("</a2ui>")
+    def has_format_content(self, content: str, *, complete: bool = False) -> bool:
+        if complete:
+            return "<a2ui" in content and "</a2ui>" in content
+        return "<a2ui" in content
 
-    is_truncated = False
-    if last_open_match:
-        last_open = last_open_match[-1].start()
-        if last_open > last_close:
-            content += "\n</a2ui>"
-            is_truncated = True
+    def unwrap(self, content: str) -> List[ResponsePart]:
+        """Unwraps/tokenizes the response content into raw Elemental HTML parts."""
+        content_lower = content.lower()
+        last_open_match = list(_A2UI_OPEN_PATTERN.finditer(content))
+        last_close = content_lower.rfind("</a2ui>")
 
-    from .compiler import TAG_PREFIX
+        is_truncated = False
+        if last_open_match:
+            last_open = last_open_match[-1].start()
+            if last_open > last_close:
+                content += "\n</a2ui>"
+                is_truncated = True
 
-    # Match <a2ui>...</a2ui>, <ui-delete-surface.../>, and <ui-call-function.../> blocks
-    block_pattern = re.compile(
-        r"<a2ui\b[^>]*>.*</a2ui>"
-        f"|<{TAG_PREFIX}delete-surface\\b[^>]*>(?:.*?</{TAG_PREFIX}delete-surface>|/>)?"
-        f"|<{TAG_PREFIX}call-function\\b[^>]*>(?:.*?</{TAG_PREFIX}call-function>|/>)?",
-        re.DOTALL | re.IGNORECASE,
-    )
-    matches = list(block_pattern.finditer(content))
+        matches = list(_BLOCK_PATTERN.finditer(content))
 
-    if not matches:
-        return [ResponsePart(text=content, a2ui_json=None)]
+        if not matches:
+            return [ResponsePart(text=content, a2ui_raw=None)]
 
-    compiler = ElementalCompiler(catalog)
-    response_parts = []
-    last_end = 0
+        response_parts = []
+        last_end = 0
 
-    for idx, match in enumerate(matches):
-        start, end = match.span()
+        for idx, match in enumerate(matches):
+            start, end = match.span()
 
-        # Clean up markdown code block wrappers around the HTML block
-        text_part = content[last_end:start]
-        text_part_stripped = re.sub(
-            r"```html\s*$", "", text_part, flags=re.IGNORECASE
-        ).strip()
+            text_part = content[last_end:start]
+            text_part_stripped = re.sub(
+                r"```html\s*$", "", text_part, flags=re.IGNORECASE
+            ).strip()
 
-        html_content = match.group(0).strip()
-        is_block_final = not (is_truncated and idx == len(matches) - 1)
+            html_content = match.group(0).strip()
+            
+            is_block_final = not (is_truncated and idx == len(matches) - 1)
+            if not is_block_final:
+                self._truncated_blocks.add(html_content)
 
-        try:
-            compiled_json = compiler.compile(
-                html_content, surface_id=surface_id, is_final=is_block_final
-            )
             response_parts.append(
                 ResponsePart(
                     text=text_part_stripped if text_part_stripped else None,
-                    a2ui_json=[compiled_json],
+                    a2ui_raw=html_content,
                 )
             )
-        except Exception:
-            # Graceful fallback: treat malformed/unparseable blocks as plain text
-            fallback_text = html_content
-            full_text = f"{text_part}\n{fallback_text}" if text_part else fallback_text
-            response_parts.append(ResponsePart(text=full_text, a2ui_json=None))
+            last_end = end
 
-        last_end = end
+        trailing_text = content[last_end:].strip()
+        if trailing_text:
+            response_parts.append(ResponsePart(text=trailing_text, a2ui_raw=None))
 
-    trailing_text = content[last_end:]
-    trailing_text_stripped = re.sub(r"^\s*```", "", trailing_text).strip()
-    if trailing_text_stripped:
-        response_parts.append(ResponsePart(text=trailing_text_stripped, a2ui_json=None))
+        return response_parts
 
-    return response_parts
+    def compile(self, format_content: str) -> List[dict[str, Any]]:
+        """Compiles raw Elemental HTML to structured A2UI messages."""
+        compiler = ElementalCompiler(self.catalog)
+        is_final = format_content not in self._truncated_blocks
+        compiled_json = compiler.compile(
+            format_content, surface_id=self.surface_id, is_final=is_final
+        )
+        self._truncated_blocks.discard(format_content)
+        return [compiled_json]
+
+    def process_chunk(self, chunk: str) -> List[ResponsePart]:
+        """Elemental is parsed as a whole HTML5 document; streaming is not supported."""
+        raise NotImplementedError(
+            "Streaming parsing is not supported for Elemental HTML."
+        )
