+++ agent_sdks/python/a2ui_agent/src/a2ui/inference_formats/experimental/elemental/prompt_generator.py
@@ -18,6 +18,7 @@ Translates standard JSON catalog schemas into TypeScript/TSX interface
 definitions and instruction blocks for on-device models.
 """
 
+import json
 import re
 from typing import Any, Optional, TYPE_CHECKING
 from a2ui.schema.catalog import A2uiCatalog
@@ -26,6 +27,30 @@ from a2ui.prompt import PromptGenerator
 
 if TYPE_CHECKING:
     from .format import ElementalFormat
+
+
+ELEMENTAL_RULES = r"""# A2UI Elemental Output Contract
+
+You must output the user interface using A2UI Elemental HTML5-like markup.
+Surround the entire output with `<body>` and `</body>` tags, including a `<link rel="catalog" href="[CATALOG_ID]">` at the start.
+**CRITICAL**: DO NOT output raw JSON or `<a2ui-json>`. Direct JSON outputs are strictly prohibited.
+
+## HTML5 Markup Rules
+
+1. **Component Tags**: Use elements prefixed with `ui-` in kebab-case (e.g. `<ui-card>`).
+2. **Component IDs**: Provide a unique `id` attribute for every component. The single top-level element MUST have `id="root"`.
+3. **Attributes**: Pass static string values as regular attributes (`variant="primary"`). Wrap numbers, booleans, and expressions in double-quoted curly braces: `elevation="{4}"`, `disabled="{true}"`.
+4. **Data Binding**: Bind data using curly braces prefixed with `$`: `value="{$/user/name}"` (absolute) or `value="{$name}"` (relative in list templates). Use `{$/items/0}` for arrays, never brackets.
+5. **Expressions**: Call catalog functions inside curly braces using named arguments: `text="{formatCurrency(value: $/price, currency: 'USD')}"`.
+6. **Slots & Children**: Nest children inside parent elements. Use the `slot` attribute to specify child properties: `<ui-card slot="leading">`.
+7. **Complex Properties**: For objects/arrays, use `<script type="application/json" slot="prop">`. For HTML/long text, use `<script type="text/html" slot="prop">`.
+8. **Templates**: For dynamic lists, nest child elements inside a `<template>` tag, and specify the bound data array path via the `path` attribute on the list component itself (e.g. `<ui-list path="{$/items}"><template>...</template></ui-list>`).
+9. **Actions**: Use `on-<property-name>` in kebab-case (e.g. `onclick="{Event('name', {args})}"`). If submitting or validating data, pass the data paths inside the event context dict (e.g. `onclick="{Event('login', {username: $/login/username})}"`).
+10. **Standalone Directives**:
+    - Data Initialization: `<script type="application/json">{"data"}</script>` at root of body.
+    - Surface Deletion: `<ui-delete-surface surface-id="id" />`.
+    - Standalone Function Call: `<ui-call-function id="id" name="func"><script type="application/json" slot="args">{"args"}</script></ui-call-function>`.
+"""
 
 
 def _schema_allows_databinding(prop_schema: Any) -> bool:
@@ -292,6 +317,52 @@ class ElementalPromptGenerator(PromptGenerator):
 
         return "\n".join(declarations)
 
+    def _replace_json_block(self, match: re.Match[str]) -> str:
+        json_content = match.group(1).strip()
+        try:
+            parsed = json.loads(json_content)
+            if isinstance(parsed, dict):
+                messages = [parsed]
+            elif isinstance(parsed, list):
+                messages = parsed
+            else:
+                return str(match.group(0))
+
+            blocks = []
+            for msg in messages:
+                if isinstance(msg, dict) and any(
+                    k in msg
+                    for k in [
+                        "createSurface",
+                        "updateDataModel",
+                        "deleteSurface",
+                        "callFunction",
+                    ]
+                ):
+                    decompiled = self._format.decompile(msg)
+                    blocks.append(decompiled)
+                else:
+                    return str(match.group(0))
+
+            return self._format.wrap_decompiled_blocks(blocks)
+        except Exception:
+            return str(match.group(0))
+
+    def transform_examples(self, raw_examples_markdown: str) -> str:
+        """Transforms JSON blocks in raw markdown into Elemental HTML syntax."""
+        if not self.catalog:
+            return raw_examples_markdown
+
+        triple_backticks = chr(96) * 3
+        pattern = rf"{triple_backticks}json\s*\n(.*?)\n{triple_backticks}"
+
+        return re.sub(
+            pattern,
+            self._replace_json_block,
+            raw_examples_markdown,
+            flags=re.DOTALL,
+        )
+
     def generate(
         self,
         role_description: str,
@@ -342,10 +413,11 @@ class ElementalPromptGenerator(PromptGenerator):
         )
 
         parts = [role_description]
-        workflow = self._format.format_description(
-            custom_workflow_description=workflow_description, catalog_id=self.catalog_id
-        )
-        parts.append(f"## Workflow Description:\n{workflow}")
+        
+        rules = ELEMENTAL_RULES.replace("[CATALOG_ID]", self.catalog_id)
+        if workflow_description:
+            rules += f"\n\n{workflow_description}"
+        parts.append(f"## Workflow Description:\n{rules}")
 
         if ui_description:
             parts.append(f"## UI Description:\n{ui_description}")
@@ -358,7 +430,7 @@ class ElementalPromptGenerator(PromptGenerator):
                 self._format.examples_path, validate=validate_examples
             )
             if raw_examples:
-                formatted_examples = self._format.transform_examples(raw_examples)
+                formatted_examples = self.transform_examples(raw_examples)
                 parts.append(f"### Examples:\n{formatted_examples}")
 
         return "\n\n".join(parts)
