+++ agent_sdks/python/a2ui_agent/src/a2ui/inference_formats/experimental/express/__init__.py
@@ -21,14 +21,13 @@ into standard A2UI v1.0 wire JSON messages and vice-versa.
 from .compiler import ExpressCompiler
 from .decompiler import ExpressDecompiler
 from .constants import SurfaceOperation
-from .parser import parse_express_response
-from .format import ExpressFormat, ExpressParser
+from .format import ExpressFormat
+from .parser import ExpressParser
 
 __all__ = [
     "ExpressCompiler",
     "ExpressDecompiler",
     "SurfaceOperation",
-    "parse_express_response",
     "ExpressFormat",
     "ExpressParser",
 ]
