+++ agent_sdks/python/a2ui_agent/src/a2ui/inference_formats/experimental/express/parser.py
@@ -19,73 +19,78 @@ from typing import Any, List, Union
 from a2ui.core.catalog import Catalog
 from a2ui.schema.catalog import A2uiCatalog
 from a2ui.parser.response_part import ResponsePart
+from a2ui.parser.parser import Parser
+from google.adk.utils.feature_decorator import experimental
 from .compiler import ExpressCompiler
 
 _A2UI_DSL_BLOCK_PATTERN = re.compile(r"<a2ui>(.*?)</a2ui>", re.DOTALL)
 
 
-def parse_express_response(
-    content: str,
-    catalog: Union[Catalog[Any, Any], A2uiCatalog],
-    surface_id: str = "main",
-) -> List[ResponsePart]:
-    """Parses response containing A2UI Express DSL and compiles it to ResponseParts.
+@experimental
+class ExpressParser(Parser):
+    """Concrete parser implementation for A2UI Express DSL responses."""
 
-    NOTE: This parser supports unclosed tag auto-closing for real-time streaming preview
-    rendering. If the final <a2ui> block is unclosed (truncated), it will be auto-closed
-    and compiled with is_final=False to discard any trailing incomplete statements.
-    IMPORTANT: For stateful continuations, client applications must accumulate streaming chunks
-    at the string level before parsing, rather than parsing chunks in isolation.
+    def __init__(self, catalog: Union[Catalog[Any, Any], A2uiCatalog], surface_id: str = "main"):
+        self.catalog = catalog
+        self.surface_id = surface_id
+        self._truncated_blocks = set()
 
-    Args:
-        content: The raw LLM response.
-        catalog: A Catalog or an A2uiCatalog.
-        surface_id: The target surface ID.
+    def has_format_content(self, content: str, *, complete: bool = False) -> bool:
+        if complete:
+            return "<a2ui>" in content and "</a2ui>" in content
+        return "<a2ui" in content
 
-    Returns:
-        A list of ResponsePart objects containing compiled JSON payload list.
-    """
-    is_truncated = False
-    last_open = content.rfind("<a2ui>")
-    last_close = content.rfind("</a2ui>")
-    if last_open != -1 and last_open > last_close:
-        content += "</a2ui>"
-        is_truncated = True
+    def unwrap(self, content: str) -> List[ResponsePart]:
+        """Unwraps/tokenizes the response content into raw Express DSL parts."""
+        # Handle unclosed tag auto-closing
+        last_open = content.rfind("<a2ui>")
+        last_close = content.rfind("</a2ui>")
+        is_truncated = False
+        if last_open != -1 and last_open > last_close:
+            content += "</a2ui>"
+            is_truncated = True
 
-    matches = list(_A2UI_DSL_BLOCK_PATTERN.finditer(content))
-    if not matches:
-        return [ResponsePart(text=content, a2ui_json=None)]
+        matches = list(_A2UI_DSL_BLOCK_PATTERN.finditer(content))
+        if not matches:
+            return [ResponsePart(text=content, a2ui_raw=None)]
 
-    compiler = ExpressCompiler(catalog)
-    response_parts = []
-    last_end = 0
+        response_parts = []
+        last_end = 0
 
-    for idx, match in enumerate(matches):
-        start, end = match.span()
-        text_part = content[last_end:start].strip()
+        for idx, match in enumerate(matches):
+            start, end = match.span()
+            text_part = content[last_end:start].strip()
 
-        dsl_content = match.group(1).strip()
-        is_block_final = not (is_truncated and idx == len(matches) - 1)
+            dsl_content = match.group(1).strip()
+            
+            is_block_final = not (is_truncated and idx == len(matches) - 1)
+            if not is_block_final:
+                self._truncated_blocks.add(dsl_content)
 
-        try:
-            compiled_json = compiler.compile(
-                dsl_content, surface_id=surface_id, is_final=is_block_final
-            )
             response_parts.append(
                 ResponsePart(
-                    text=text_part if text_part else None, a2ui_json=[compiled_json]
+                    text=text_part if text_part else None,
+                    a2ui_raw=dsl_content,
                 )
             )
-        except Exception:
-            # Graceful fallback: treat malformed/unparseable blocks as plain text so the app doesn't crash
-            fallback_text = f"<a2ui>\n{dsl_content}\n</a2ui>"
-            full_text = f"{text_part}\n{fallback_text}" if text_part else fallback_text
-            response_parts.append(ResponsePart(text=full_text, a2ui_json=None))
+            last_end = end
 
-        last_end = end
+        trailing_text = content[last_end:].strip()
+        if trailing_text:
+            response_parts.append(ResponsePart(text=trailing_text, a2ui_raw=None))
 
-    trailing_text = content[last_end:].strip()
-    if trailing_text:
-        response_parts.append(ResponsePart(text=trailing_text, a2ui_json=None))
+        return response_parts
 
-    return response_parts
+    def compile(self, format_content: str) -> List[dict[str, Any]]:
+        """Compiles raw Express DSL to structured A2UI messages."""
+        compiler = ExpressCompiler(self.catalog)
+        is_final = format_content not in self._truncated_blocks
+        compiled_json = compiler.compile(
+            format_content, surface_id=self.surface_id, is_final=is_final
+        )
+        self._truncated_blocks.discard(format_content)
+        return [compiled_json]
+
+    def process_chunk(self, chunk: str) -> List[ResponsePart]:
+        """Express DSL is parsed as a whole script block; streaming is not supported."""
+        raise NotImplementedError("Streaming parsing is not supported for Express DSL.")
